Extract bot core models and rules
