+++ src/bot/__init__.py
@@ -1,0 +1,2 @@
+"""Plain bot models, rules, and services."""
+
